requirements.txt
absl-py==1.4.0
altair==5.1.2
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
beautifulsoup4==4.12.2
blinker==1.7.0
brewer2mpl==1.4.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
cmake==3.29.2
colorama==0.4.6
comm==0.2.2
constantly==15.1.0
contourpy==1.1.0
cryptography==40.0.1
cssselect2==0.7.0
cv==1.0.0
cycler==0.10.0
daphne==4.0.0
debugpy==1.8.1
decorator==4.4.2
Django==3.0.5
django-ckeditor==6.5.1
django-cleanup==5.1.0
django-crispy-forms==1.9.2
django-js-asset==2.0.0
django-widget-tweaks==1.4.8
djangorestframework==3.15.1
easyocr==1.7.0
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.13.4
filterpy==1.4.5
Flask==3.0.3
flatbuffers==23.5.26
fonttools==4.41.1
frozendict==2.3.8
fsspec==2024.3.1
gast==0.4.0
ggplot==0.11.5
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
h11==0.14.0
h5py==3.9.0
html5lib==1.1
httplib2==0.22.0
hydra-core==1.3.2
hyperlink==21.0.0
idna==3.4
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
incremental==22.10.0
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==2.13.1
kiwisolver==1.4.4
lazy_loader==0.4
libclang==16.0.6
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.5
multitasking==0.0.11
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.74
openpyxl==3.1.2
opt-einsum==3.3.0
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==23.1
pandas==2.0.2
pandas-datareader==0.10.0
parso==0.8.4
patsy==0.5.3
peewee==3.17.0
Pillow==10.0.0
platformdirs==4.2.1
plotly==5.18.0
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.8
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post5
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.16.1
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
pyOpenSSL==23.1.0
pyparsing==3.0.9
pypdf==4.2.0
pypi==2.1
pypng==0.20220715.0
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-magic==0.4.27
pytube==15.0.0
pytz==2023.3
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.2
qrcode==7.4.2
redis==4.5.3
referencing==0.31.0
regex==2023.10.3
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.0
roboflow==1.1.27
robotframework==7.0
robotframework-pythonlibcore==4.4.1
robotframework-seleniumlibrary==6.2.0
rpds-py==0.13.0
rsa==4.9
scikit-image==0.23.2
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
selenium==4.14.0
service-identity==21.1.0
shapely==2.0.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.4.1
sqlparse==0.4.3
stack-data==0.6.3
statsmodels==0.14.0
stop-words==2018.7.23
streamlit==1.28.2
svglib==1.5.1
sympy==1.12
tbb==2021.12.0
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
textblob==0.17.1
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2024.4.24
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
torch==2.3.0
torchvision==0.18.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.3
trio==0.22.2
trio-websocket==0.11.1
Twisted==22.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.2
ultralytics==8.0.114
uritools==4.0.2
urllib3==1.26.16
validators==0.22.0
watchdog==3.0.0
wavio==0.0.7
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.2
wordcloud==1.9.3
wrapt==1.15.0
wsproto==1.2.0
xhtml2pdf==0.2.15
yfinance==0.2.31
zipp==3.17.0
zope.interface==6.0
